Create README.md following modified project template
This commit creates the README.md file in this repository, following a
modified version of the Software Project Template provided by NASA SLIM.
The template has been modified to better fit the purpose of the
repository, which is not a software project. This may not be the best
practice, but it serves as a reference for future attempts at creating
a standardized template for non-software projects.

Currently, approximately 40% of the modified template has been completed,
with unfinished content sections marked as (NOT DONE). This commit serves
as an opportunity to create a standardized README.md file and may provide
a clear presentation for potential readers to understand the purpose of
the repository.

Future updates to the repository will follow the Atomic Approach, with
each completed block being committed separately.